+++ tests/Support/Types/MyCustomScalarString.php
@@ -7,11 +7,13 @@ namespace Rebing\GraphQL\Tests\Support\Types;
 use Exception;
 use GraphQL\Error\Error;
 use GraphQL\Language\AST\Node;
+use GraphQL\Type\Definition\Type;
 use GraphQL\Error\InvariantViolation;
 use GraphQL\Type\Definition\ScalarType;
 use GraphQL\Language\AST\StringValueNode;
+use Rebing\GraphQL\Support\Contracts\TypeConvertible;
 
-class MyCustomScalarString extends ScalarType
+class MyCustomScalarString extends ScalarType implements TypeConvertible
 {
     /**
      * Serializes an internal value to include in a response.
@@ -28,7 +30,7 @@ class MyCustomScalarString extends ScalarType
     }
 
     /**
-     * Parses an externally provided value (query variable) to use as an input
+     * Parses an externally provided value (query variable) to use as an input.
      *
      * In the case of an invalid value this method must throw an Exception
      *
@@ -44,7 +46,7 @@ class MyCustomScalarString extends ScalarType
     }
 
     /**
-     * Parses an externally provided literal value (hardcoded in GraphQL query) to use as an input
+     * Parses an externally provided literal value (hardcoded in GraphQL query) to use as an input.
      *
      * In the case of an invalid node or value this method must throw an Exception
      *
@@ -57,10 +59,15 @@ class MyCustomScalarString extends ScalarType
      */
     public function parseLiteral($valueNode, ?array $variables = null)
     {
-        if (!$valueNode instanceof StringValueNode) {
-            throw new InvariantViolation('Expected node of type ' . StringValueNode::class . ' , got ' . get_class($valueNode));
+        if (! $valueNode instanceof StringValueNode) {
+            throw new InvariantViolation('Expected node of type '.StringValueNode::class.' , got '.get_class($valueNode));
         }
 
         return $valueNode->value;
+    }
+
+    public function toType(): Type
+    {
+        return new static();
     }
 }